refactor(general): Refactor Fix Develop
