+++ app/src/main/java/com/example/nuberjam/utils/NetworkConnectionInterceptor.kt
@@ -1,4 +1,4 @@
-package com.example.nuberjam.utils.network
+package com.example.nuberjam.utils
 
 import android.content.Context
 import android.net.ConnectivityManager
