+++ app/src/main/java/com/example/nuberjam/utils/NoConnectivityException.kt
@@ -1,4 +1,4 @@
-package com.example.nuberjam.utils.network
+package com.example.nuberjam.utils
 
 import java.io.IOException
