--- app/src/main/java/com/example/nuberjam/utils/extension/GeneralExt.kt
@@ -1,28 +1,0 @@
-package com.example.nuberjam.utils.extension
-
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.LifecycleOwner
-import androidx.lifecycle.lifecycleScope
-import androidx.lifecycle.repeatOnLifecycle
-import kotlinx.coroutines.launch
-import okhttp3.MediaType.Companion.toMediaType
-import okhttp3.MultipartBody
-import okhttp3.RequestBody.Companion.asRequestBody
-import java.io.File
-
-fun File.toMultipartBody(field: String): MultipartBody.Part {
-    val requestImageFile = this.asRequestBody("image/*".toMediaType())
-    return MultipartBody.Part.createFormData(
-        field,
-        this.name,
-        requestImageFile
-    )
-}
-
-fun LifecycleOwner.launchLifecycleScopeOnStarted(block: suspend () -> Unit) {
-    lifecycleScope.launch {
-        repeatOnLifecycle(Lifecycle.State.STARTED) {
-            block()
-        }
-    }
-}
